requirements.txt
pandas
selenium
webdriver-manager
python-dotenv
openpyxl
streamlit
plotly
pydantic-settings
python-dotenv
